+++ android/src/main/kotlin/com/nankai/flutter_nearby_connections/EndpointStatus.kt
@@ -1,0 +1,6 @@
+package com.nankai.flutter_nearby_connections
+
+class EndpointStatus {
+    var name: String? = null
+    var status: String? = null
+}